Fix indentation and typo in the filename
